began prepping name module for v0.1.0-dev
