--- src/lib/modules/generator/CMakeLists.txt
@@ -1,16 +1,0 @@
-set(NG_SOURCES 
-		${CMAKE_CURRENT_SOURCE_DIR}/name-generator.cpp
-		${CMAKE_CURRENT_SOURCE_DIR}/generator.cpp
-		${CMAKE_SOURCE_DIR}/src/lib/utils.cpp
-)
-
-set(NG_INCLUDES
-		${CMAKE_CURRENT_SOURCE_DIR}/generator.h
-)
-
-set(CMAKE_CXX_FLAGS "-g -Wall --std=c++11")
-
-add_executable(name-generator ${NG_SOURCES})
-
-install(TARGETS name-generator DESTINATION bin)
-install(FILES ${NG_INCLUDE} DESTINATION lib)
