+++ src/lib/modules/names/CMakeLists.txt
@@ -1,0 +1,16 @@
+set(NG_SOURCES 
+		${CMAKE_CURRENT_SOURCE_DIR}/name-generator.cpp
+		${CMAKE_CURRENT_SOURCE_DIR}/names.cpp
+		${CMAKE_SOURCE_DIR}/src/lib/utils.cpp
+)
+
+set(NG_INCLUDES
+		${CMAKE_CURRENT_SOURCE_DIR}/names.h
+)
+
+set(CMAKE_CXX_FLAGS "-g -Wall --std=c++11")
+
+add_executable(name-generator ${NG_SOURCES})
+
+install(TARGETS name-generator DESTINATION bin)
+install(FILES ${NG_INCLUDE} DESTINATION lib)
