+++ src/lib/utils.h
@@ -21,5 +21,6 @@ std::string rightpad(std::string str, int len = 0, char ch = ' ');
 std::istream& safeGetline(std::istream& is, std::string& t);
 bool print_file(std::string file);
 int verbose(std::string log = "", int status = -1);
+int random(int min, int max);
 
 #endif /* SRC_UTILS_H_ */
