+++ src/tests/generator-test.cpp
@@ -7,7 +7,7 @@ This is free software: you are free to change and redistribute it.
 There is NO WARRANTY, to the extent permitted by law.
 */
 #include <string>
-#include "generator.h"
+#include "names.h"
 
 int main(int argc, char* argv[]) {
     std::string race = argv[1];
